Update SCENARIOS.txt
Updated all of the scenarios for the user stories. There are some that seem kind of old or redundant, so I included parts of them in more important scenarios or just left them with a comment. If anyone wants to go through them and present a different idea as to how the thing should work, don't be afraid to make any changes! I'm not even completely sure how some of the pages are working currently or how you guys want your specific pages to work so this is all just how I think things will end up working or how I think people have been implementing.